Fix order dependent spec failure in exec specs
The provider/exec/windows_spec.rb spec was creating exec resources
explicitly using the Windows provider as the provider. The Windows
provider was being evaluated as suitable and thus the default provider
for the type. Then in future specs, it could be used since it was the
default, even though it wasn't necessarily suitable, causing failures.
Now, we just clear the defaultprovider on the exec type after all the
specs in the file have finished, ensuring it will be recalculated
appropriately when needed.

Reviewed-By: Matt Robinson <matt@puppetlabs.com>