Fix #7004 - plural for indirection ending with 'e' is broken
This is not a complete fix.
This got broken when the status indirection was added around
d1f1858ea52d3089fd2088994e80d6f49d7e0347.

The problem is that the regex matching is greedy so any indirection
ending in e (and pluralized in 'es') will lose its trailing 'e'.

Since all this was done for the status indirection, my fix is to
treat differently this indirection only, instead of trying to fix
the general case (and breaking it).

Signed-off-by: Brice Figureau <brice-puppet@daysofwonder.com>