Make string_to_sid_ptr block optional
Previously the Puppet::Util::Windows::Security#string_to_sid_ptr
method required a block to be passed to it.

This commit makes the block optional, and simply returns true if the
string SID, e.g. 'S-1-5-18', can be converted into a SID pointer
(pointer to the binary SID structure). If the string SID cannot be
converted, then a Puppet::Util::Windows::Error will be thrown. If the
string SID is invalid, then the P::U::W::E#code will return 1337.

This commit is being implemented so that we can validate SIDs prior to
converting from string SIDs to names.