Remove duplicate SID resolution code
Previously the User and Group classes defined their own method for
resolving a name into a SID with the assumption that we might need to
customize the WMI query for each type of object (by including the
SIDType in the where clause). But it turns out that the Name and
Domain properties of the Win32_Account table are a compound key, so
the method for resolving names into SIDs can be refactored.

Also, eliminated duplicate, and slightly incorrect code from the
Windows security integration test (it wasn't restricting the query to
local accounts only).