(#9461) 'puppet resource package' fails on Windows
This commit corrects a basic failure to enumerate package resources
on Windows. In previous versions, 'package' enumeration failed due to
the provider subsystem's inability to exec any process neccesary to
do the enumerations.

The general fix expands the Puppet::Util.which command when running
on Windows to search for EXE variants if the target is not found.
For example, a search for 'gem' fails, because 'gem' (like all other
Ruby and Puppet binaries) is wrapped in a batch file, creating 'gem.bat'.

If a lookup fails on Windows, and the passed value doesn't contain a file
extension, 'which' will examine the value of the environment 'PATHEXT'
setting, which contains a delimited list of known extensions. 'which'
will append each extension to the value (in order of appearance) and
reevaluate. If no PATHEXT value exists, the code defaiults to the
basic Windows binary extensions (.bat;.ps1;.exe;.com). The first match
found wins, and the translated value is cached against future lookups.

Spec tests have been updated to reflect this change.