Maint: Fix the "provider" parameter documentation
The provider parameter had a chunk of code that seemed to be designed to
include the name of the type; however, it has always been broken and was
resulting in bizarre English.

This commit changes the way we introspect the current type -- it's hacky, but
it gives us a correct answer -- and adds two tests to the type spec to ensure
this documentation is being generated correctly.

Paired-with: Matt Robinson <matt@puppetlabs.com>