Move group management into providers
Move provider-specific logic from the file group property to the posix
file provider. For example, mapping group IDs to names and vice-versa
is provider specific, but detecting if the property value is insync?
is common across all file providers.

Modify the windows file provider to allow getting and setting the
group via its 'group' and 'group=' methods. Also, on Windows the file
owner can be a group, and the file group can be a user, so the same
method for mapping names and SIDs can be used for both owner and
group.

Added test cases where there weren't any.