(#9775) Only list managed resources in the resources file
In ticket #8667 puppet was made to write out a list of resources in the
catalog. However, it sometimes wrote out resources that weren’t actually
being managed, like audited resources.

Turns out that there was already a method on the resources to ask if
they were managed, it it did what I needed, so I used it.