(#9326) Support plaintext passwords in Windows 'user' provider.
This commit adds password support to the Windows 'user' provider.
When setting the password, the resource value is passed unmolested
to the underlying ADSI 'SetPassword' method.  During a get,
the password is checked using the Windows 'LoginUser' function with
tha current user name and password. If the call succeeds, the password
is returned; if unsuccessful, ':absent' is returned.

The spec test has been updated to reflect these changes.