(#9328) Retrieve user and group SIDs on windows.
This commit implements the ID get-method for Windows 'user' and
'group' providers. Prior to this commit, 'puppet resource {user,group}'
would not return an id property for each resource. With this change,
the Windows user/group security identifier (SID) is provided as the id.

The uid and gid properties are read only. Setting these values is still
not supported for Windows providers, and attempting to set either
property will result in a failure message. Attempting to set the
user 'gid' or 'shell' properties will likewise fail.

Spec tests have been updated.