(#9837) Start the cleanup of the puppet resource application
In order to use the same logic for the puppet resource application in
the puppetral MCollective agent it helps to figure out what logic in the
resource app is there for what reason.

Specs that declare an object responds to a method are redundant when the
method is called later to test the output.

Also have tests that assert errors are raised, also assert the messages
associated with those errors.

Reviewed-by: Max Martin <max@puppetlabs.com>