(#9796) ssh_authorized_key supports whitespace again
Remove the value validation for the name parameter.

1c7f0c3530 caused a regression: puppet does not support whitespace in
the name anymore, e.g. if you specify

    ssh_authorized_key { 'fancy user':
      ensure => present,
      type   => rsa,
      key    => 'AAAAB3...iopMxe5aWw==',
    }

you will get the error "Resourcename must not contain whitespace."

While spaces are in general used as a delimiter OpenSSH does support
these in the comment field and the provider did already handle
whitespace correctly. So the type should not forbid it.