(#6352) Caching Fact values for Puppet's use
The previous version was loading all facts for
retrieval of a single value, and this caches the entire
set in memory, reloading them whenver Facts.load
is called.

An alternative approach is to extend the accessor method
('[]') through to the facter terminus.  Given that this
way Puppet has control over the caching and such, this seemed
a better approach.

Signed-off-by: Luke Kanies <luke@puppetlabs.com>