(#9832) 2.7.4 StoreConfigs regression with PostgreSQL.
The extraction of StoreConfigs to a terminus caused a regression for
PostgreSQL users, in that we had the wrong case for the title of resources in
the database when matching.

This wasn't caught because MySQL and SQLite do case insensitive matching, and
our stub data was wrongly using a downcased name to test.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>