(#9837) Split parameter pruning from manifest formatting
The code we really want to use in the MCollective puppetral agent is
just the pruning of parameters, not the formatting of the resource to a
manifest text.  Therefore, we're splitting that out so pruning and
manifestification can be called separately.

This also adds tests for the parameter pruning behavior.

Reviewed-by: Max Martin <max@puppetlabs.com>