(#9832) Test failures with some ActiveRecord versions.
The previous PostgreSQL fix had issues with some versions of ActiveRecord
causing test failures; this seems to be tied to specific versions of both AR
and SQLite triggering the failure mode.

The root cause was a badly stubbed record in the database, which is corrected
by this change, and the problem resolved.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>