(#9838) Return the tranaction report when doing a ral save
When using puppet resource from the command line, using `puppet
resource` to do a save will log error messages to the console when
saving using the ral indirection.  However, this doesn't help when using
that indirection in Ruby like you might from MCollective's puppetral
agent.

So we now return the transaction report you get from applying the
catalog.

The only place we could find this indirection being used was in the
`puppet resource` application, although it's possible that code external
to puppet uses this indirection and will need to change what it expects
for the return value of save.

Reviewed-by: Max Martin <max@puppetlabs.com>