Undo change to failing test on 1.8.5
One of the file integration specs was failing because Tempfile does
not produce random filenames. As a result, a test was attempting to
create a directory, but failing, because a file already exists from a
previous test.

This commit changes the test back to the way it was -- it creates a
directory in which it creates files and directories.