(#8411) Fix change group for POSIX file provider
During the file type refactor, the POSIX file provider was
accidentally changed to invoke the File.chgrp and File.lchgrp
methods. These methods don't actually exist and this was causing
acceptance test failures.

This commit changes the POSIX file provider to invoke the File.chown
and File.lchown methods when changing the group. Updated the spec
tests, and also added integration tests, which simply change the owner
and group to the current values. While not the best test, it will work
in non-root contexts and will ensure this doesn't happen again.