Change test to not call 'rm -rf'
Previously the test was failing on Windows because there were problems
with local file serving on Windows. That part was fixed, so the only
remaining issue was changing the test to not call 'rm -rf'.

This commit converts the tests to the new style, using tmpfile,
etc. and removes the fails_on_windows tag.