Add Windows-specific tests when user parameter specified in exec
Previously, the tests were marked as fails_on_windows, because there
wasn't an exec provider on Windows. Now that we have an exec provider,
the tests were failing, because we do not support executing processes
as a different user on Windows (as that requires a password). Since
the behavior is different on Windows vs POSIX, I added tests for the
Windows behavior.