Stub method for getting roles from the user provider
Previously, these tests were failing because the Windows user provider
does not implements methods for getting the user's roles. These tests
actually pass on other platforms that don't support
'manages_solaris_rbac', because the Puppet::Provider::NameService
autogenerates default methods for providers that don't implement
getter and setter properties for valid properties.

Since this test is testing the ability for the user to autorequire its
roles (as opposed to testing roles), this commit stubs the roles
method on the provider, enabling the test to run on Windows.