(#9158) Support old and new versions of STOMP gem.
The STOMP gem, up to release 1.1.4, used `send` to push messages to the
queue.  From 1.1.5 onward this is replaced by `publish`, and `send` issues a
deprecation warning, but still works.

This commit dynamically adapts to the best available method: if `publish` is
present we use that, otherwise we fall back to `send`, and fail cleanly if we
drift further.

This should behave correctly with all known versions of the STOMP gem, avoid
deprecation warnings, and still end up with the right results.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>