(#9832) General StoreConfigs regression.
The previous fixes around PostgreSQL were not complete in addressing the
regression: some StoreConfigs exported and imported resources would not be
found.

This removes the last bits that were causing regressions, and additional
testing shows that the generated SQL is now identical between 2.7.3 and the
current version.

This should resolve all the remaining regressions.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>