(#2848) Set `certdnsnames` values into the CSR.
We were incorrectly applying the `certdnsnames` setting when a certificate was
signed, not when a CSR was generated.  This would misapply the server setting
to all nodes in the infrastructure, rather than applying the client-supplied
version.

This commit adds `certdnsnames` to the CSR, allowing the client to pass their
desired set of names to the CA to be propagated into the final certificate.

It does not alter the behaviour when the certificate is generated.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>