(#9831) Standardize Windows provider confining
The newly created providers for Windows were using a combination of
':operatingsystem => :windows' and ':feature => :microsoft_windows'
for their confine and defaultfor calls.  This brings them in line to
all use the ':operatingsystem => :windows' form.

The general reasoning being that :operatingsystem is used for things
that are platform specific, where :feature is used for things that are
supported across multiple platforms, but require that external (to
Puppet or the base platform) support exist on the host.