(#2848) List subject alt names in output of puppet cert --list
Given that it's potentially risky to sign a certificate which has alt
names, we now surface that information so the user can make an informed
decision when signing. This also includes some changes to the format of
the output, grouping certs by status to make it more easily scannable.

Conflicts:

	lib/puppet/ssl/certificate_authority/interface.rb
	spec/unit/ssl/certificate_authority/interface_spec.rb

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>