(#2848) extract the subjectAltName value from the CSR.
When a CSR is submitted, we need to inspect it and determine if there are any
subjectAltNames present in the request, and if those are legal.

The logic to validate the content of an extension is complex, and decomposing
the content requires digging into various X509 and ASN.1 related classes.

This commit adds a `subject_alt_names` method that will extract the extensions
containing subjectAltName values from the CSR, and reformat them into a useful
set of content suitable for use in Ruby.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>

Conflicts:

	lib/puppet/ssl/certificate_request.rb
	spec/unit/ssl/certificate_request_spec.rb

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>