(#7296) Make the Debian service provider handle services that don't conform to the debain policy manual.
This change is to support initscripts that do not support the --query
method of invoke-rc.d used by the Debian provider to determine if
service is enabled.

The fix checks that the link count in /etc/rc?.d is equal to 4, which is
the number of links that should be present when using the Debian service
provider, which is done by `update-rc.d #{service} defaults`.