(#2848) Rename `certdnsnames` to match new behaviour.
As part of the update to SSL the behaviour of the `certdnsnames` options
changed sufficiently that it would be terribly confusing to retain it.

Instead, modify the setting to warn that it is ignored, and add a new setting
to set the default subjectAltName value for bootstrapping a master
certificate.

This retains the one really useful part of the feature, without the risk that
someone will accidentally use the old name and receive a nasty surprise.

It should also draw more attention to formerly insecure configurations.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>

Conflicts:

	acceptance/pending/ticket_3360_allow_duplicate_csr_with_option_set.rb
	acceptance/tests/ticket_5477_master_not_dectect_sitepp.rb
	lib/puppet/defaults.rb
	lib/puppet/ssl/host.rb
	spec/integration/defaults_spec.rb
	spec/unit/ssl/host_spec.rb
	test/lib/puppettest/exetest.rb
	test/network/server/webrick.rb
	test/ral/type/filesources.rb

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>