(#2848) Rename `certdnsnames` to match new behaviour.
As part of the update to SSL the behaviour of the `certdnsnames` options
changed sufficiently that it would be terribly confusing to retain it.

Instead, modify the setting to warn that it is ignored, and add a new setting
to set the default subjectAltName value for bootstrapping a master
certificate.

This retains the one really useful part of the feature, without the risk that
someone will accidentally use the old name and receive a nasty surprise.

It should also draw more attention to formerly insecure configurations.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>