Wire up the `setbycli` slot in Puppet settings.
There was an existing slot on the setting object that tracked if the value had
been set on the command line or not, but it was not wired up to the rest of
the system.

This wires that up, allowing us to use that to determine if the
`dns-alt-names` setting should apply just to the local node, or to all nodes,
safely from the `puppet cert` application.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>