(#2848) Migrate `dns-alt-names` back to settings.
Now that all the other changes have been made, the `dns-alt-names` can migrate
back to the settings where they apply only for generating the CSR for the
local machine.

The one exception to that rule is the `puppet cert generate` application, in
which we want to pass any *command line* value of `dns-alt-names` into the CSR
for a foreign node, but not the value from anywhere but the command line.

Since the settings now track the CLI origin of a setting, we can use that to
derive the data and pass it around appropriately to make that so.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>