(#2848) Don't strip the subjectAltName label when listing.
The X.509 `subjectAltName` field contains a set of labeled values; each label
defines a namespace that the rest of the content fits within.  DNS is one of
those, and is the only one that Puppet actually uses, but there are plenty of
others including email addresses, IP, a DN, or a URI.

When we displayed the values we used to strip that label off, which hid
potentially critical information from the user.  Instead we should preserve,
and display, that data.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>