(#2848) Eliminate redundant `master_dns_alt_names`.
Now that `dns_alt_names` applies to local CSR generation, there is no need
for a special configuration option only applied to a master certificate.

Eliminating that option allows us to simplify the overall model, and provide
more uniform access to the `subjectAltName` setting when required.

Documentation about the option is also updated.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>