Add support for dns-alt-names option to `puppet certificate generate`
This option has the same behavior as it does with `puppet cert
generate`. That is, if the option is set via the CLI, it will always be
set on the CSR. If it is not set via the CLI, but is supplied in the
config file, it will only be set on the CSR if the CSR is for the
current host.