(#9983) Read file content in binary mode when backing up
Previously, Puppet::FileBucket::Dipper#backup would read a file's
content in text mode, which on Windows causes '\r\n' to be translated
to '\n'. As a result, the backup method was calculating the wrong
checksum for the to-be-backed-up file.

This commit changes the backup method to open the file in binary mode,
which is a noop on Unix.